On apple gcc v 5465 and above c99 is enforced by -std=gnu99 the same as for gcc 4.3 and above. The problem is, one cannot use either extern inline or extern __inline__ __attribute__ ((__gnu_inline__)) or static inline for the inlined functions in gmp.h. The first can't be used since it isn't c99, the second can't be used since the attribute is not supported by apple's gcc and the third cannot be used since this makes the library functions also static inline which causes duplicates.
The only solution which works for now is to not allow inlined functions in gmp.h. This obviously results in a performance deficit. But better a slow mpir than none at all!!

Ironically the first solution I came up with for the c99 problem was actually more robust than what we have here, but it was terribly inelegant.

This current patch works on Jason's home core2. Who knows if it breaks everything else!