Added core2, x86_64 as supported build targets and core2 as a supported CPU.
Changed Slowaris to Solaris.

Added some information about the MSVC project files.

Corrected a typo.

Remove the word "unsupported" in the infomation about the macos port.

Changed a handful of GMP's to MPIR's and added my name to the list of author's.

Changed the copyright to mention that MPIR started with GMP v.4.2.1 which was copyright FSF, blah, blah, blah.